+++ internal/server/core/ports.go
@@ -3,17 +3,9 @@ package core
 import (
 	"context"
 
-	shcore "github.com/ghazlabs/idn-remote-entry/internal/shared/core"
+	"github.com/ghazlabs/idn-remote-entry/internal/shared/core"
 )
 
-type Storage interface {
-	Save(ctx context.Context, v shcore.Vacancy) (*shcore.VacancyRecord, error)
-}
-
-type VacancyResolver interface {
-	Resolve(ctx context.Context, url string) (*shcore.Vacancy, error)
-}
-
-type Notifier interface {
-	Notify(ctx context.Context, v shcore.VacancyRecord) error
+type Queue interface {
+	Put(ctx context.Context, req core.SubmitRequest) error
 }
