+++ internal/shared/core/error.go
