+++ internal/vacancy-worker/driven/resolver/hqloc/model.go
