+++ internal/vacancy-worker/driven/resolver/model.go
