+++ internal/vacancy-worker/driven/resolver/util/util.go
