+++ internal/vacancy-worker/driven/resolver/vacancy.go
