+++ internal/vacancy-worker/driven/storage/model.go
